+++ zenml/core/backends/orchestrator/aws/orchestrator_aws_backend.py
@@ -18,9 +18,8 @@ import time
 from typing import Text, Dict, Any
 
 from zenml.core.backends.orchestrator.aws import utils
-from zenml.core.backends.orchestrator.base.orchestrator_base_backend import \
-    OrchestratorBaseBackend
-from zenml.core.repo.repo import Repository
+from zenml.core.backends.orchestrator import OrchestratorBaseBackend
+from zenml.core.repo import Repository
 from zenml.core.standards import standard_keys as keys
 from zenml.utils import path_utils
 from zenml.utils.constants import ZENML_BASE_IMAGE_NAME
