+++ zenml/core/backends/orchestrator/beam/orchestrator_beam_backend.py
@@ -16,8 +16,7 @@
 import os
 from typing import Any, Text, Dict
 
-from zenml.core.backends.orchestrator.base.orchestrator_base_backend import \
-    OrchestratorBaseBackend
+from zenml.core.backends.orchestrator import OrchestratorBaseBackend
 from zenml.core.backends.orchestrator.beam.zenml_beam_orchestrator import \
     ZenMLBeamlDagRunner
 from zenml.core.pipelines.utils import parse_yaml_beam_args
