+++ zenml/core/backends/orchestrator/gcp/orchestrator_gcp_backend.py
@@ -23,15 +23,14 @@ from typing import Text
 import googleapiclient.discovery
 from google.oauth2 import service_account as sa
 
-from zenml.core.backends.orchestrator.base.orchestrator_base_backend import \
-    OrchestratorBaseBackend
-from zenml.core.repo.repo import Repository
+from zenml.core.backends.orchestrator import OrchestratorBaseBackend
+from zenml.core.repo import Repository
 from zenml.core.standards import standard_keys as keys
 from zenml.utils import path_utils
+from zenml.utils import requirement_utils
 from zenml.utils.constants import ZENML_BASE_IMAGE_NAME, \
     ZENML_TRAINER_IMAGE_NAME, GCP_ENTRYPOINT
 from zenml.utils.logger import get_logger
-from zenml.utils import requirement_utils
 
 requirement_utils.check_integration(requirement_utils.GCP_INTEGRATION)
