+++ zenml/core/backends/orchestrator/kubernetes/orchestrator_kubernetes_backend.py
@@ -23,9 +23,8 @@ from kubernetes import client as k8s_client
 from kubernetes import config as k8s_config
 from kubernetes.config.config_exception import ConfigException
 
-from zenml.core.backends.orchestrator.base.orchestrator_base_backend import \
-    OrchestratorBaseBackend
-from zenml.core.repo.repo import Repository
+from zenml.core.backends.orchestrator import OrchestratorBaseBackend
+from zenml.core.repo import Repository
 from zenml.core.standards import standard_keys as keys
 from zenml.utils import path_utils
 from zenml.utils.constants import ZENML_BASE_IMAGE_NAME, K8S_ENTRYPOINT
