+++ zenml/core/backends/processing/processing_base_backend.py
@@ -15,7 +15,7 @@
 
 from typing import Optional, List, Text
 
-from zenml.core.backends.base_backend import BaseBackend
+from zenml.core.backends import BaseBackend
 
 
 class ProcessingBaseBackend(BaseBackend):
