+++ zenml/core/backends/processing/processing_spark_backend.py
@@ -13,8 +13,7 @@
 #  permissions and limitations under the License.
 """Definition of the Spark Processing Backend"""
 
-from zenml.core.backends.processing.processing_base_backend import \
-    ProcessingBaseBackend
+from zenml.core.backends.processing import ProcessingBaseBackend
 
 
 class ProcessingSparkBackend(ProcessingBaseBackend):
